before fl_chart

+++ lib/models/the1G.dart
@@ -5,12 +5,12 @@ class The1G {
 
   });
 
-  int d;
+  double d;
   double c;
 
 
   factory The1G.fromJson(Map<String, dynamic> json) => The1G(
-    d: json["d"],
+    d: json["d"].toDouble(),
     c: json["c"].toDouble(),
 
   );
